Fixed rendering issue with the smoke trail particle effect
The issue was that when viewed from the side, the smoke trail would not render and would disappear. Fixed this by unchecking "Use Fixed Relative Bounding Box" under Bounds in the Details panel. Note that this is not a universal solution as this could be a performance hit on some systems as the bounding box will be calculated every single frame.